requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
appnope==0.1.3
asttokens==2.0.5
async-timeout==4.0.2
attrs==21.4.0
backcall==0.2.0
black==22.3.0
charset-normalizer==2.0.12
click==8.1.3
decorator==5.1.1
executing==0.8.3
flake8==4.0.1
frozenlist==1.3.0
idna==3.3
ipdb==0.13.9
ipython==8.4.0
iso8601==1.0.2
jedi==0.18.1
matplotlib-inline==0.1.3
mccabe==0.6.1
multidict==6.0.2
mypy-extensions==0.4.3
parso==0.8.3
pathspec==0.9.0
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==2.5.2
prompt-toolkit==3.0.29
ptyprocess==0.7.0
pure-eval==0.2.2
pycodestyle==2.8.0
pyflakes==2.4.0
Pygments==2.12.0
python-dotenv==0.20.0
six==1.16.0
stack-data==0.3.0
toml==0.10.2
tomli==2.0.1
traitlets==5.2.2.post1
twitchio==2.3.0
typing_extensions==4.2.0
wcwidth==0.2.5
yarl==1.7.2
